requirements.txt
asgiref==3.9.2
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
click==8.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.1
daphne==4.1.0
Django==5.2.7
django-cors-headers==4.3.1
django-debug-toolbar==6.0.0
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
joblib==1.5.2
msgpack==1.1.1
nltk==3.8.1
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
qrcode==8.0
redis==5.0.1
regex==2025.9.18
scikit-learn==1.5.0
scipy==1.13.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
TextBlob==0.18.0
threadpoolctl==3.6.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
zope.interface==8.0.1